refactor(router): extract screen routes from RegistrationRouter (WP6.0.1)
- Create routes/ folder with 7 route components
- Extract SilentAssignRoute (17 lines)
- Extract ClearCourtRoute (43 lines)
- Extract HomeRoute (92 lines)
- Extract AdminRoute (110 lines)
- Extract SuccessRoute (169 lines)
- Extract GroupRoute (177 lines)
- Extract CourtRoute (297 lines) - largest with court selection logic
- Create routes/index.js barrel export
- Update RegistrationRouter to delegate via props spread

RegistrationRouter: 851 → 59 lines (-792 lines, -93%)
Route components total: 905 lines

Verbatim JSX extraction. No behavior change.
