+++ src/registration/router/routes/index.js
@@ -1,0 +1,7 @@
+export { SuccessRoute } from './SuccessRoute.jsx';
+export { SilentAssignRoute } from './SilentAssignRoute.jsx';
+export { HomeRoute } from './HomeRoute.jsx';
+export { AdminRoute } from './AdminRoute.jsx';
+export { GroupRoute } from './GroupRoute.jsx';
+export { CourtRoute } from './CourtRoute.jsx';
+export { ClearCourtRoute } from './ClearCourtRoute.jsx';
